Add "Warning" title
Added a TitleManager and a flashing red "Warning" title for Fright
Night.
Fright Night will have to be redone - the server will be on permanent
night mode, so the timer will have to do on its own.
Perhaps take out the day counter and have Fright Night trigger for 60
seconds every 20 minutes?